+++ resources/components/laraknife/panels/edit.blade.php
@@ -1,7 +1,7 @@
 @props(['title' => 'Edit', 'error' => ''])
 <div id="main-content" class="container mt-5">
     <x-laraknife.panels.common title="{{ $title }}" />
-    <fieldset class="lkn-edit-panel">
+    <fieldset class="lkn-panel">
         {{ $slot }}
         <div class="row lkn-empty-line-above">
             <x-laraknife.buttons.store width1="2" width2="4" />