+++ resources/components/laraknife/panels/menu.blade.php
@@ -1,7 +1,7 @@
 @props(['title'])
 <div id="main-content" class="container mt-5">
     <x-laraknife.panels.common title="{{ $title }}" />
-    <fieldset class="lkn-create-panel">
+    <fieldset class="lkn-panel">
         {{ $slot }}
     </fieldset>
 </div>