+++ resources/components/laraknife/panels/noform.blade.php
@@ -1,7 +1,7 @@
 @props(['title' => 'Edit', 'error' => '', 'class' => ''])
 <div id="main-content" class="container mt-5">
     <x-laraknife.panels.common title="{{ $title }}" />
-    <fieldset class="lkn-noform-panel {{$class}}">
+    <fieldset class="lkn-panel {{$class}}">
         {{ $slot }}
     </fieldset>
 </div>